requirements.txt
Flask==2.2.2
Flask_Migrate==4.0.0
Flask_SQLAlchemy==2.5.1
Pillow==9.4.0
PyQt5==5.15.7
PyQt6==6.4.0
requests==2.28.1
RS_Image_Convertor==0.3.1
save_thread_result==0.0.9
